Call upsertPoll from post and comment callbacks, rather than revisions
