+++ packages/lesswrong/server/collections/comments/mutations.ts
@@ -6,7 +6,7 @@ import { accessFilterSingle } from "@/lib/utils/schemaUtils";
 import { userCanDo, userOwns } from "@/lib/vulcan-users/permissions";
 import { addReferrerToComment, assignPostVersion, checkCommentForSpamWithAkismet, checkModGPTOnCommentCreate, checkModGPTOnCommentUpdate, commentsAlignmentEdit, commentsAlignmentNew, commentsEditSoftDeleteCallback, commentsNewNotifications, commentsNewOperations, commentsNewUserApprovedStatus, commentsPublishedNotifications, createShortformPost, handleForumEventMetadataEdit, handleForumEventMetadataNew, handleReplyToAnswer, invalidatePostOnCommentCreate, invalidatePostOnCommentUpdate, lwCommentsNewUpvoteOwnComment, moveToAnswers, newCommentsEmptyCheck, newCommentsPollResponseCheck, newCommentsRateLimit, newCommentTriggerReview, setTopLevelCommentId, trackCommentRateLimitHit, updatedCommentMaybeTriggerReview, updateDescendentCommentCountsOnCreate, updateDescendentCommentCountsOnEdit, updatePostLastCommentPromotedAt, updateUserNotesOnCommentRejection, validateDeleteOperations } from "@/server/callbacks/commentCallbackFunctions";
 import { updateCountOfReferencesOnOtherCollectionsAfterCreate, updateCountOfReferencesOnOtherCollectionsAfterUpdate } from "@/server/callbacks/countOfReferenceCallbacks";
-import { upsertPolls } from "@/server/callbacks/revisionCallbacks";
+import { upsertPolls } from "@/server/callbacks/forumEventCallbacks";
 import { sendAlignmentSubmissionApprovalNotifications } from "@/server/callbacks/sharedCallbackFunctions";
 import { createInitialRevisionsForEditableFields, reuploadImagesIfEditableFieldsChanged, uploadImagesInEditableFields, notifyUsersOfNewPingbackMentions, createRevisionsForEditableFields, updateRevisionsDocumentIds, notifyUsersOfPingbackMentions } from "@/server/editor/make_editable_callbacks";
 import { logFieldChanges } from "@/server/fieldChanges";
@@ -87,13 +87,17 @@ export async function createComment({ data }: CreateCommentInput, context: Resol
     newDoc: documentWithId,
     props: afterCreateProperties,
   });
-  const completePollUpsert = async () => context.repos.forumEvents.ensurePostId({ commentId: documentWithId._id });
-  await completePollUpsert();
 
   documentWithId = await notifyUsersOfPingbackMentions({
     newDoc: documentWithId,
     props: afterCreateProperties,
   });
+
+  await upsertPolls({
+    revisionId: documentWithId.contents_latest,
+    comment: documentWithId,
+    context,
+  })
 
   await updateCountOfReferencesOnOtherCollectionsAfterCreate('Comments', documentWithId);
 
@@ -163,6 +167,12 @@ export async function updateComment({ selector, data }: UpdateCommentInput, cont
     newDoc: updatedDocument,
     props: updateCallbackProperties,
   });
+
+  await upsertPolls({
+    revisionId: updatedDocument.contents_latest,
+    comment: updatedDocument,
+    context,
+  })
 
   await updateCountOfReferencesOnOtherCollectionsAfterUpdate('Comments', updatedDocument, oldDocument);
